(#6) Split the file `src/GRUPRecord.h`
Split the content of the file `src/GRUPRecord.h` into three pieces and added
each corresponding part to the existing `src/game/<game>/GRUPRecord.h` that
was already in existance.

After that the original file was empty, so it was deleted.

Also the `src/game/<game>/GRUPRecord.h` files now required explicit inclusion
of `Allocator.h` since the content that was moved to them depended on the
class template declared there.