(#6) Cleaned `TES4Record` of Skyrim dependencies
Moved Skyrim's specific String functionality that was located in the
generic master record (TES4Record) to the Skyrim ModFile implementation,
as it's only used from records belonging to Skyrim's implementation.

For that I had to break a cyclic dependence on TES5Record <-> TES5File
by introducing a new zero sized abstract class TES5ModFile that only
publishes the method required to access the StringsLookup
implementation.

As a result of this I can remove the inclusion of SkyrimCommon.h
from `TES4Record.cpp`